Consolidate reactive search multilists into a wrapper component
